Deleted original zipped NPI file and replaced it
This new zipped file has all the files and documentation that comes with the download of a monthly NPI replacement file. This data is as of June 8th and was downloaded on June 16th. The README explains how large the raw data file is and tells users to unzip on their local machine.